backend/requirements-ocr.txt
fastapi
uvicorn
python-multipart
transformers
torch
torchvision
einops
tiktoken
pillow
accelerate
# flash-attn  # Note: Requires compilation on some systems, recommend install manually if needed
